Harden deploy staging shell discovery
